earthquakes sphinx-gallery examples
