+++ src/geovista/cache.py
@@ -39,7 +39,7 @@ TextureLike = str | pv.Texture
 BASE_URL: str = "https://github.com/bjlittle/geovista-data/raw/{version}/data/"
 
 #: Pin to use the specific geovista-data repository version for geovista resources.
-DATA_VERSION: str = "2023.12.11"
+DATA_VERSION: str = "2023.12.12"
 
 #: Environment variable to override pooch cache manager path.
 ENV: str = "GEOVISTA_CACHEDIR"
