+++ src/geovista/examples/scalar_data/README.rst
@@ -1,0 +1,2 @@
+Scalar Data
+===========
