+++ src/geovista/examples/scalar_data/__init__.py
@@ -1,0 +1,24 @@
+# Copyright (c) 2021, GeoVista Contributors.
+#
+# This file is part of GeoVista and is distributed under the 3-Clause BSD license.
+# See the LICENSE file in the package root directory for licensing details.
+
+"""The examples showcase various geovista features and capabilities.
+
+Each example is importable and runnable as a standalone script from the command
+line.
+
+Note that, the examples are installed as part of the geovista package
+and can be easily accessed through the geovista command line interface
+(CLI) entry-point.
+
+.. code: bash
+
+  # For further details refer to the geovista examples (CLI) help.
+  geovista examples --help
+
+Notes
+-----
+.. versionadded:: 0.5.0
+
+"""
