+++ src/geovista/registry.txt
@@ -35,8 +35,6 @@ tests/images/clouds.png 227b833d01d82b30cb9c016e3652c4b3ef4196f9368befbc2f118eea
 tests/images/clouds_robin.png f522db3d4a5739eeebd33fdf46363c38ea1b3027aea0470e175d1992d84ab658
 tests/images/curvilinear__from_2d__orca_moll.png 9b98f0d912ff38b8c1ca4a4e3329cf1001d7132069487d07bed083c22e2b5c4d
 tests/images/curvilinear__from_2d__orca.png 8552e2010bb058f190694cc7da3d096988edb30406b2d9ad1d64165eaa2f836a
-tests/images/earthquakes.png 2995a9277dfbc0505a6e175a5affa418a066ad39482c3de98678fe8521d56aa6
-tests/images/earthquakes_wink1.png bb929dc99d2ea390a217929a9217fc7b1e7828639cd9a0fcc51484b342b88d6f
 tests/images/point_cloud__from_points__orca_cloud_eqc.png 6421665bc4bff1128b2876b7b83978b3c6e894be607b39764e136f013b795ddc
 tests/images/point_cloud__from_points__orca_cloud.png a5fa2a43eb6f8b1ab2adf6f6e757830b07305a87a3bb9abf64c48d2d2d8a066b
 tests/images/rectilinear__from_1d__oisst.png 92da94eb490eaa72acca8938e22edacf530b43a15bd215633bb6111d5380bfba
@@ -49,6 +47,8 @@ tests/images/rectilinear__from_2d__synthetic_face_m1n1.png 9ef84bc7a41d237063f81
 tests/images/rectilinear__from_2d__synthetic_face_m1n1_robin.png c23335fd376d81bf47f9d9f70059fe0f03fd52b243426ea693e65ef9195340fd
 tests/images/rectilinear__from_2d__synthetic_node_m1n1_moll.png d93e9907bd306b95c2e67b5ed99466f08edb3cdd35de85bb4026076422f9b5e3
 tests/images/rectilinear__from_2d__synthetic_node_m1n1.png 9d00923e48dad39ba03d903b18a5151ffb085b98e8eba4ad0601ad8630aef5b8
+tests/images/scalar_data__earthquakes.png 2995a9277dfbc0505a6e175a5affa418a066ad39482c3de98678fe8521d56aa6
+tests/images/scalar_data__earthquakes_wink1.png bb929dc99d2ea390a217929a9217fc7b1e7828639cd9a0fcc51484b342b88d6f
 tests/images/spatial_index__uber_h3.png 08cb9cbb5f3ae41d1955efe7cdcfac999b2c6e67072f57871dde85ab49205289
 tests/images/unstructured__dynamico.png ba4f2f4a77f065597a9f2db3bf08d7308df7dec49916c973e55961394ed9916c
 tests/images/unstructured__dynamico_poly.png 1473440458fdda137fee6d1b9c1cd57d12cc7b098e130667bfaa8faddf25e22f
